define 'baremetal' libvirt network for cluster com
Up until now, we were using the 'default' libvirt network for the
cluster communication. The issue with that is that its definition can
differ from env to env and also that touching it could mess with
pre-existing libvirt domains (VMs).

The new baremetal network aims to simulate the datacenter network that
we'll have to interact with, with its DNS entries for the API and DNS
VIPs.